knowledge_reinforcer/requirements.txt
requests
readability-lxml
youtube-transcript-api
markdownify
PyYAML
